fix memory request for very large images CCBUGS: 420868 CCBUGS: 416266

+++ core/libs/dimg/dimg_data.cpp
@@ -166,7 +166,9 @@ void DImg::copyImageData(const QExplicitlySharedDataPointer<Private>& src)
 
 size_t DImg::allocateData() const
 {
-    size_t size  = m_priv->width * m_priv->height * (m_priv->sixteenBit ? 8 : 4);
+    size_t size  = (quint64)m_priv->width  *
+                   (quint64)m_priv->height *
+                   (quint64)(m_priv->sixteenBit ? 8 : 4);
     m_priv->data = DImgLoader::new_failureTolerant(size);
 
     if (!m_priv->data)
